AssertionForge/requirements.txt
absl-py==2.1.0
aioconsole==0.8.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
anytree==2.12.1
argcomplete==3.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.5
async-timeout==4.0.3
asyncer==0.0.7
asyncio-atexit==1.0.1
attrs==23.2.0
autograd==1.6.2
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.4.0
azure-storage-blob==12.20.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.1.3
beartype==0.18.5
beautifulsoup4==4.12.3
bidict==0.23.1
black==25.1.0
bleach==6.2.0
blis==0.7.11
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
cffi==1.16.0
chainlit==2.0.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.5
chromadb==0.5.4
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
cobble==0.1.4
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coolname==2.2.0
cramjam==2.8.3
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.8
dask==2024.7.0
dask-expr==1.1.7
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
devtools==0.12.2
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docutils==0.21.2
easyocr==1.7.2
email_validator==2.2.0
environs==11.0.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
evaluate==0.4.2
exceptiongroup==1.2.1
executing==2.0.1
Faker==37.1.0
fastapi==0.115.6
fastapi-cli==0.0.4
fastjsonschema==2.21.1
fastparquet==2024.5.0
filetype==1.2.0
FLAML==2.1.2
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gensim==4.3.2
git-filter-repo==2.47.0
graphviz==0.20.3
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.0
humanfriendly==10.0
hvac==2.3.0
hyppo==0.4.0
imageio==2.36.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
jiter==0.5.0
joblib==1.4.2
json5==0.11.0
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
klepto==0.2.5
kubernetes==30.1.0
lancedb==0.9.0
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langcodes==3.4.0
langsmith==0.3.8
language_data==1.2.0
Lazify==0.4.0
lazy_loader==0.4
linkify-it-py==2.0.3
llama-cloud==0.1.7
llama-index==0.12.9
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.9
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.12
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.2
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.19
llvmlite==0.43.0
locket==1.0.0
lxml==5.3.0
mammoth==1.9.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
markdownify==0.13.1
marko==2.1.2
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.1.3
mmh3==4.1.0
monotonic==1.6
msal==1.29.0
msal-extensions==1.2.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
ninja==1.11.1.2
nltk==3.9.1
notebook_shim==0.2.4
numba==0.60.0
oauthlib==3.2.2
ollama==0.2.1
onnxruntime==1.16.3
openai
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pathvalidate==3.2.3
patsy==0.5.6
pdfminer.six==20240706
pexpect==4.9.0
platformdirs==4.2.2
ply==3.11
portalocker==2.10.0
posthog==3.5.0
POT==0.9.4
pox==0.3.4
preshed==3.0.9
prettytable==3.10.2
prometheus_client==0.21.1
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.24.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
py==1.11.0
py_sv_parser==0.3.0
pyaml-env==1.2.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen>=0.2.27
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-settings==2.6.1
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.13.0
PyNaCl==1.5.0
pynndescent==0.5.13
pyparsing==3.1.2
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproject_hooks==1.1.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==3.3.0
python-multipart==0.0.18
python-pptx==1.0.2
python-socketio==5.12.1
pytz==2024.1
pyverilog
pyzmq==26.3.0
ratelimit
ratelimiter
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge_score
rpds-py==0.19.0
rsa==4.9
Rtree==1.3.0
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.0.1
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.2
striprtf==0.0.26
swifter
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
terminado==0.18.1
textual==0.70.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
together==1.2.2
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
uptrace==1.28.2
uri-template==1.3.0
uritemplate==4.1.1
uvicorn==0.25.0
uvloop==0.19.0
vcdvcd==2.3.6
wasabi==1.1.3
watchfiles==0.20.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.16.0
wsproto==1.2.0
xgboost
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.18.3
zipp==3.19.2
zstandard==0.23.0
IPython
ruamel.yaml>=0.17
